+++ CHANGELOG.md
@@ -945,6 +945,7 @@
 - [Use `numpy` & co. from Enso!][7678]
 - [Changed layout of local libraries directory, making it easier to reference
   projects next to each other][7634]
+- [Support runtime checks of intersection types][7769]
 - [Merge `Small_Integer` and `Big_Integer` types][7636]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
@@ -1083,6 +1084,7 @@
 [7613]: https://github.com/enso-org/enso/pull/7613
 [7678]: https://github.com/enso-org/enso/pull/7678
 [7634]: https://github.com/enso-org/enso/pull/7634
+[7769]: https://github.com/enso-org/enso/pull/7769
 [7636]: https://github.com/enso-org/enso/pull/7636
 
 # Enso 2.0.0-alpha.18 (2021-10-12)